backend/requirements.txt
asgiref==3.9.1
Brotli==1.1.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
colorama==0.4.6
cssselect2==0.8.0
Django==5.2.5
django-anymail==13.0.1
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
fonttools==4.59.1
idna==3.10
mysqlclient==2.2.7
pillow==11.3.0
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
PyJWT==2.10.1
pyotp==2.9.0
pyphen==0.17.2
python-decouple==3.8
qrcode==8.2
requests==2.32.4
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
tzdata==2025.2
urllib3==2.5.0
weasyprint==66.0
webencodings==0.5.1
zopfli==0.2.3.post1
